% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nuclear.R
\name{nuclear_least_squares}
\alias{nuclear_least_squares}
\title{Mutivariate least-squares with nuclear norm penalty}
\usage{
nuclear_least_squares(a, b, t = 1, x = rnorm(dim(a)[2]) \%*\%
  t(rnorm(dim(b)[2])), w = 0, ridge_penalty = 0, max_its = 10)
}
\arguments{
\item{a}{The n x p matrix A}

\item{b}{The n x q matrix B}

\item{t}{The nuclear norm constraint}

\item{x}{Intitialization}

\item{w}{Initial weight}

\item{ridge_penalty}{Ridge penalty}

\item{max_its}{Maximum iterations}
}
\description{
Minimizes ||AX - B||^2_F + ridge_penalty * ||X||^2_F subject to nuclear norm(X) < t
}
\examples{
a <- matrix(rnorm(20 * 10), 20, 10)
res <- svd(matrix(rnorm(10 * 5), 10, 5))
x0 <- res$u \%*\% diag(c(.8, .6, .3, .2, .1)) \%*\% t(res$v)
b <- a \%*\% x0 + 0.01 * matrix(rnorm(20 * 5), 20, 5)
x_hat <- nuclear_least_squares(a, b, 2, max_its = 20)
sum((x_hat - x0)^2)
}

